Moved hadoop/hdfs/hadoop_jar to luigi.contrib
The old paths still work for now, but will throw a warning that it's deprecated

Also added copyright headers in hadoop_jar
